Fix failing CodeClimate checks
Something has changed in the CodeClimate platform (update of the used
PHP version?) that breaks the CodeClimate tests.

### Change used PHPCodeSniffer version

The first error is:

> PHP Deprecated:  Array and string offset access syntax with curly
> braces is deprecated in
> /usr/src/app/vendor/squizlabs/php_codesniffer/CodeSniffer.php
> on line x
```

CodeClimate uses by default version 2.8.x
(see https://docs.codeclimate.com/docs/phpcodesniffer). The 3.6.x
version is available using the "beta channel".

Fix this by adding the channel to your `.codeclimate.yml` file:

```yaml
plugins:
  phpcodesniffer:
    enabled: true
    channel: beta
    config:
      standard: ".phpcs.xml"
```

### Change in .phpcs.xml

The `.phpcs.xml` file is fetched from this repository and added to the
project before CodeClimate starts running the checks and plugins. A new
error emerges after changing the PHPCodeSniffer version to 3.6.x:

> Exception: ERROR: Referenced sniff
> "./vendor/drupal/coder/coder_sniffer/Drupal" does not exist.

This is fixed in the `config/phpcs.xml` file that is fetched from this
repository. CodeClimate has support for the Drupal & DrupalPractice rule
sets so we don't need to refer to the local version:

```xml
<?xml version="1.0" encoding="UTF-8"?>
<ruleset>
  <description>PHP CodeSniffer ruleset for Drupal.</description>

  <!-- Minified files don't have to comply with coding standards. -->
  <exclude-pattern>*.min.css</exclude-pattern>
  <exclude-pattern>*.min.js</exclude-pattern>

  <rule ref="Drupal"/>
  <rule ref="DrupalPractice"></rule>

  <!-- PSR-12 sniffs -->
  <rule ref="PSR12.Files.DeclareStatement"/>
</ruleset>
```